data: split album.js
