+++ src/data/things/album/index.js
@@ -1,0 +1,2 @@
+export * from './Album.js';
+export * from './TrackSection.js';
